requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
mysqlclient>=2.2.0
pymysql>=1.1.0
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
pytest>=7.4.0
pytest-django>=4.5.0
pytest-allure-adaptor>=2.12.0
allure-pytest>=2.13.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
uvicorn>=0.24.0
